Refactor: Make StoreItems.swift
